+++ gatsby/package.json
@@ -43,7 +43,7 @@
     "prismjs": "^1.15.0",
     "react": "^16.8.4",
     "react-dom": "^16.8.4",
-    "react-helmet": "^5.2.0",
+    "react-helmet": "^5.2.1",
     "react-markdown": "^4.0.8",
     "remark-github-plugin": "^1.3.1",
     "remove": "^0.1.5",
