+++ gatsby/src/pages/search.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link, graphql } from 'gatsby';
-
+import SearchResults from '../components/searchResults';
 import Layout from '../layout/layout';
 import CallToAction from '../layout/call-to-action';
 import TopicsGrid from '../layout/topics-grid';
@@ -24,21 +24,9 @@ class Index extends React.Component {
 							<div className="container-fluid">
 								<div className="row">
 									<div className="col-md-8 search-results">
-{/*
-									<script>
-										function parseParamsFromUrl() {
-										var queryString = window.location.search;
-											queryString = queryString.substring(11);
-											queryString = queryString.split("+").join(" ");
-										return queryString;
-										}
-										var urlParams = parseParamsFromUrl();
-										document.getElementById('piodocsearch').setAttribute('value', urlParams);
-   								   </script>
 
-*/}
-									<div id="addsearch-results"></div>
 
+									<SearchResults />
